Clean up dataId sorting.
